Move ReasonsForRejection to top-level namespace
