+++ app/components/provider_interface/status_box_components/rejected_component.html.erb
@@ -7,7 +7,7 @@
   <div class="govuk-inset-text">
     <%= render ReasonsForRejectionComponent.new(
       application_choice: @application_choice,
-      reasons_for_rejection: ProviderInterface::ReasonsForRejection.new(@application_choice.structured_rejection_reasons),
+      reasons_for_rejection: ReasonsForRejection.new(@application_choice.structured_rejection_reasons),
     ) %>
   </div>
 <% else %>
